Changed Path.restrict() to not take an Option
The advantages of explicitly supplying Some or None do not really
outweight the limitations of using an Option (compiler unable to deduce
the type of None), which constrain the inputs to Path or PathBuf. As it
is very easy to pass "." or "", which will both mean the current working
directory and hence are equivalent to None, it seems sensible to remove
the Option, which allows use of AsRef. This then supports a wider range
of inputs, including &str, String, and OsString.