Moved ip!() macro to rubedo_macros crate
Moved the ip!() macro to the new rubedo_macros crate in preparation for
it being rewritten into a proc_macro, as a crate with proc_macros is not
allowed to provide anything else. The proc_macros crate is considered a
subcrate of this crate, and is not intended to be used on its own.