Added chrono::MonthsExt
The additions in this commit are dependent upon as-yet unreleased Chrono
functionality, which will arrive in Chrono 0.4.32. Specifically, this PR
has been merged, allowing inspection of the internal quantity of Months:

https://github.com/chronotope/chrono/pull/1373

Due to the Chrono maintainers preferring Months.as_u32(), this commit
adds Duration-style convenience methods as an external extension to the
chrono::Months type.

  - Added num_months() and num_years() to extend chrono::Months.

  - Added tests and documentation.